requirements.txt
# requirements.txt is a pip requirements file which describes the necessary
# distributions required to run this project.

--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/
--extra-index-url=http://dist.pinaxproject.com/fresh-start/

Django==1.3.1
Pinax
psycopg2

django-debug-toolbar==0.9.1
django-staticfiles==1.2.1
django_compressor==1.1.2
django-friendship

django-mailer==0.2a1
django-email-confirmation==0.2
django-timezones==0.2
pytz==2011n
django-openid==0.3a1
python-openid==2.2.5
metron==0.1
django_forms_bootstrap
django-announcements==0.1.2
django-pagination==1.0.7
django-notification==0.2
idios==0.2a1

















